add blank lines; use @class

+++ IMFoundation/FZMessage.h
@@ -1,6 +1,8 @@
 @interface FZMessage : NSObject
-@property unsigned long long flags;
+
+@property NSUInteger flags;
 @property (nonatomic, retain) NSString *handle;
 @property (nonatomic, retain) NSString *sender;
 @property (nonatomic, retain) NSDate *timeRead;
+
 @end
